+++ src/zope/minmax/_minmax.py
@@ -5,6 +5,11 @@ from zope.minmax import interfaces
 
 @zope.interface.implementer(interfaces.IAbstractValue)
 class AbstractValue(persistent.Persistent):
+    """
+    Abstract implementation of `zope.minmax.interfaces.IAbstractValue`.
+
+    Subclasses *must* implement `_p_resolveConflict`.
+    """
 
     def __init__(self, value=None):
         self.value = value
@@ -22,6 +27,11 @@ class AbstractValue(persistent.Persistent):
     __nonzero__ = __bool__
 
     def _p_resolveConflict(self, old, commited, new):
+        """
+        Subclasses must implement this method.
+
+        :raises NotImplementedError: Unless subclasses override.
+        """
         raise NotImplementedError()
 
 class Maximum(AbstractValue):